Reorganized code.
